feat: opt out of space layout when on entity page
We now use layout groups in Next.js to opt out of the `/space/[id]` layout when rendering entity pages. This is so we can share the layout between `/space/[id]` subroutes (like the Governance tab) without automatically opting-in all other child routes.
