+++ apps/web/app/space/(entity)/[id]/[entityId]/activity/layout.tsx
