+++ apps/web/app/space/(entity)/[id]/[entityId]/default-entity-page.tsx
