+++ apps/web/app/space/(entity)/[id]/[entityId]/layout.tsx
