+++ apps/web/app/space/(entity)/[id]/[entityId]/page.tsx
