+++ apps/web/app/space/(entity)/[id]/[entityId]/triples/page.tsx
